requirements.txt
streamlit
langchain
python-dotenv
langchain-groq
langchain-core